avm1: Move getURL to MovieClip.prototype (fix #804)
For example, _root.getURL should work.
